Change 'lamports' variables to double
The 'lamports' variables in multiple classes were previously stored as long or int data types. Change these to double data type to prevent errors arising from large numbers. This is in accordance with the Solana RPC API specification which recommends 'lamports' to be handled as double. Also, minor addition to print out the 'AccountInfo' in the 'MainnetTest'.

+++ src/main/java/org/p2p/solanaj/rpc/types/Reward.java
@@ -12,7 +12,7 @@ public class Reward {
     private String pubkey;
 
     @Json(name = "lamports")
-    private int lamports;
+    private double lamports;
 
     @Json(name = "postBalance")
     private String postBalance;
